[compiler][analysis] Renamed LivenessAnalysis to Liveness to eliminate redundant naming
